TECH: Refactor lens buttons in navigation bar
