--- src/react/features/lenses/LensesBar.js
@@ -1,7 +1,0 @@
-import React from 'react'
-import LensesBarView from './LensesBarView'
-import { useLensesBar } from './useLensesBar'
-
-const LensesBar = () => <LensesBarView {...useLensesBar()} />
-
-export default LensesBar
